- Changed: Some experimental changes to reduce the memory footprint of Game objects.  For this first modification changed regen ticks, race, and speed from int to short.  These few changes reduce the Storm footprint by 2K.  All of these may be changed to ushort but at this time I'm not sure if any of the values can be negative.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk/GameServer@2410 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ packets/Client/168/UseSpellHandler.cs
@@ -127,11 +127,11 @@ namespace DOL.GS.PacketHandler.Client.v168
 
 				if ((m_flagSpeedData & 0x200) != 0)
 				{
-					player.CurrentSpeed = -(m_flagSpeedData & 0x1ff); // backward movement
+					player.CurrentSpeed = (short)(-(m_flagSpeedData & 0x1ff)); // backward movement
 				}
 				else
 				{
-					player.CurrentSpeed = m_flagSpeedData & 0x1ff; // forward movement
+					player.CurrentSpeed = (short)(m_flagSpeedData & 0x1ff); // forward movement
 				}
 				player.IsStrafing = (m_flagSpeedData & 0x4000) != 0;
 				player.TargetInView = (m_flagSpeedData & 0xa000) != 0; // why 2 bits? that has to be figured out
