+++ spells/UnbreakableSpeedDecreaseSpellHandler.cs
@@ -128,7 +128,7 @@ namespace DOL.GS.Spells
 			GameNPC npc = owner as GameNPC;
 			if (npc != null)
 			{
-				int maxSpeed = npc.MaxSpeed;
+				short maxSpeed = npc.MaxSpeed;
 				if (npc.CurrentSpeed > maxSpeed)
 					npc.CurrentSpeed = maxSpeed;
 			}
